requirements.txt
pandas>=2.2.0
pyarrow>=16.0.0
requests>=2.31.0

boto3>=1.34.0
google-cloud-storage>=2.18.0
google-auth>=2.30.0
